Create module for grats config code/types
