+++ src/lib.ts
@@ -21,32 +21,10 @@ import { GratsDefinitionNode, TypeContext } from "./TypeContext";
 import { validateSDL } from "graphql/validation/validate";
 import { applyServerDirectives, DIRECTIVES_AST } from "./serverDirectives";
 import { extend } from "./utils/helpers";
+import { ParsedCommandLineGrats } from "./gratsConfig";
 
 export { applyServerDirectives } from "./serverDirectives";
-
-export type ConfigOptions = {
-  // Should all fields be typed as nullable in accordance with GraphQL best practices?
-  // https://graphql.org/learn/best-practices/#nullability
-  nullableByDefault?: boolean;
-
-  // Should Grats report TypeScript type errors?
-  // Defaults to `false`.
-  reportTypeScriptTypeErrors?: boolean;
-
-  // A string to prepend to the generated schema text. Useful for copyright headers or
-  // other information to the generated file. Set to `null` to omit the default
-  // header.
-  schemaHeader?: string | null;
-
-  // Unstable do not use
-  EXPERIMENTAL_codegenPath?: string;
-};
-
-export type ParsedCommandLineGrats = ts.ParsedCommandLine & {
-  raw: {
-    grats: ConfigOptions;
-  };
-};
+export * from "./gratsConfig";
 
 // Construct a schema, using GraphQL schema language
 // Exported for tests that want to intercept diagnostic errors.
@@ -73,46 +51,6 @@ export function buildSchemaResultWithHost(
   }
 
   return ok(applyServerDirectives(schemaResult.value));
-}
-
-const DEFAULT_HEADER = `# Schema generated by Grats (https://grats.capt.dev)
-# Do not manually edit. Regenerate by running \`npx grats\`.
-# @generated`;
-
-// TODO: Make this return diagnostics
-export function validateGratsOptions(
-  options: ts.ParsedCommandLine,
-): ParsedCommandLineGrats {
-  const gratsOptions = { ...(options.raw?.grats ?? {}) };
-  if (gratsOptions.nullableByDefault === undefined) {
-    gratsOptions.nullableByDefault = true;
-  } else if (typeof gratsOptions.nullableByDefault !== "boolean") {
-    throw new Error(
-      "Grats: The Grats config option `nullableByDefault` must be a boolean if provided.",
-    );
-  }
-  if (gratsOptions.reportTypeScriptTypeErrors === undefined) {
-    gratsOptions.reportTypeScriptTypeErrors = false;
-  } else if (typeof gratsOptions.reportTypeScriptTypeErrors !== "boolean") {
-    throw new Error(
-      "Grats: The Grats config option `reportTypeScriptTypeErrors` must be a boolean if provided.",
-    );
-  }
-  if (gratsOptions.schemaHeader === undefined) {
-    gratsOptions.schemaHeader = DEFAULT_HEADER;
-  } else if (
-    typeof gratsOptions.schemaHeader !== "string" &&
-    gratsOptions.schemaHeader !== null
-  ) {
-    throw new Error(
-      "Grats: The Grats config option `schemaHeader` must be a string or `null` if provided.",
-    );
-  }
-
-  return {
-    ...options,
-    raw: { ...options.raw, grats: gratsOptions },
-  };
 }
 
 function extractSchema(
